// src/model/user.ts
export type AuthUser = {
  name: string
  username: string
  email: string
  image?: string
}

export type SimpleUser = Pick<AuthUser, 'username' | 'image'>

export type HomeUser = AuthUser & {
  following: SimpleUser[]
  followers: SimpleUser[]
  bookmarks: string[]
}

export type SearchUser = Omit<AuthUser, 'email'> & {
  following: number
  followers: number
}

export type ProfileUser = SearchUser & {
  posts: number
}